Started working on fixing the blurriness of the UI on High DPI screens.
The fix is easy (simply a corresponding setting in the manifest file), but due to a lot of hard-coded sizes in pixels, a lot of widgets do not look right when scaled.

This requires going through the code and wrapping every call to wxSize() with literals by this->FromDIP() so that the number of pixels will be scaled according to the current DPI settings.