Fixed GesturalScorePage
Allowing a line width of more than 1 pixel made the drawing very slow. This fix therefore also changes the way the lines are drawn: Instead of drawing every single segment of line (essentially calling DrawLine() for each pixel), the points on the line are gathered in a vector and then drawn all at once in a single call. This is much faster.